functions/requirements.txt
firebase-functions
firebase-admin
google-cloud-storage
google-cloud-firestore
google-generativeai
PyMuPDF
Pillow
cloudevents